gmodel

+++ app/forms.py
@@ -1,5 +1,5 @@
 from flask_wtf import FlaskForm
-from wtforms import StringField, SubmitField, DateField, DecimalField
+from wtforms import StringField, SubmitField, DateField, DecimalField, IntegerField
 from wtforms.validators import DataRequired
 from datetime import datetime
 
@@ -13,6 +13,6 @@ class RecordForm(FlaskForm):
     symbol = StringField('Symbol', validators=[DataRequired()])
     date = DateField(
         'Date', default=datetime.today, validators=[DataRequired()])
-    quantity = DecimalField('Quantity', validators=[DataRequired()])
+    quantity = IntegerField('Quantity', validators=[DataRequired()])
     price = DecimalField('Price', validators=[DataRequired()])
     submit = SubmitField('Submit')
